Refactoring: extracting FindPackageRoot
