+++ internal/packages/packages.go
@@ -1,0 +1,68 @@
+package packages
+
+import (
+	"io/ioutil"
+	"os"
+	"path/filepath"
+
+	"github.com/pkg/errors"
+	"gopkg.in/yaml.v2"
+)
+
+const PackageManifestFile = "manifest.yml"
+
+type packageManifest struct {
+	Name    string `json:"name"`
+	Type    string `json:"type"`
+	Version string `json:"version"`
+}
+
+func FindPackageRoot() (string, bool, error) {
+	workDir, err := os.Getwd()
+	if err != nil {
+		return "", false, errors.Wrap(err, "locating working directory failed")
+	}
+
+	dir := workDir
+	for dir != "." {
+		path := filepath.Join(dir, PackageManifestFile)
+		fileInfo, err := os.Stat(path)
+		if err == nil && !fileInfo.IsDir() {
+			ok, err := isPackageManifest(path)
+			if err != nil {
+				return "", false, errors.Wrapf(err, "verifying manifest file failed (path: %s)", path)
+			}
+			if ok {
+				return dir, true, nil
+			}
+		}
+
+		if dir == "/" {
+			break
+		}
+		dir = filepath.Dir(dir)
+	}
+	return "", false, nil
+}
+
+func ReadPackageManifest(path string) (*packageManifest, error) {
+	content, err := ioutil.ReadFile(path)
+	if err != nil {
+		return nil, errors.Wrapf(err, "reading file body failed (path: %s)", path)
+	}
+
+	var m packageManifest
+	err = yaml.Unmarshal(content, &m)
+	if err != nil {
+		return nil, errors.Wrapf(err, "unmarshalling package manifest failed (path: %s)", path)
+	}
+	return &m, nil
+}
+
+func isPackageManifest(path string) (bool, error) {
+	m, err := ReadPackageManifest(path)
+	if err != nil {
+		return false, errors.Wrapf(err, "reading package manifest failed (path: %s)", path)
+	}
+	return m.Type == "integration" && m.Version != "", nil // TODO add support for other package types
+}
